# Fix attnum mismatch in pg_statistic after drop column and database restore