--- a/backup/queries_statistics.go
+++ b/backup/queries_statistics.go
@@ -77,6 +77,8 @@ func GetAttributeStatistics(connectionPool *dbconn.DBConn, tables []Table, proce
 	for _, table := range tables {
 		tablenames = append(tablenames, table.FQN())
 	}
+	// In the query below we need to adjust the `staattnum` as some columns could have been dropped, but during database
+	// restore the tables will be restored from scratch, and their columns will have different attnums.
 	query := fmt.Sprintf(`
 	SELECT c.oid,
 		quote_ident(n.nspname) AS schema,
@@ -84,7 +86,7 @@ func GetAttributeStatistics(connectionPool *dbconn.DBConn, tables []Table, proce
 		quote_ident(a.attname) AS attname,
 		quote_ident(t.typname) AS type,
 		s.starelid,
-		s.staattnum,
+		s.staattnum - (SELECT COUNT(1) FROM pg_attribute WHERE attrelid = c.oid AND attnum < s.staattnum AND attisdropped = TRUE) AS staattnum,
 		%s
 		s.stanullfrac,
 		s.stawidth,