@whitehawk 2024-06-28
NOTE: If one wonders about possible performance impact of the delta, here is some execution time measurement done. On a db with 2000 tables (with 120 columns each), the backup time with statistics is following:

with the fix:
```
real	1m7,734s
user	0m25,875s
sys	0m16,997s
```
without the fix:
```
real	1m0,412s
user	0m22,166s
sys	0m14,506s
```
which doesn't look dramatic.


<details>

<summary>The code for creation test tables is following:</summary>

```
CREATE OR REPLACE FUNCTION create_many_tables()
RETURNS VOID
AS $$
BEGIN

	for counter in 1..2000 loop
		EXECUTE FORMAT($fmt$DROP TABLE IF EXISTS test_table_%1$s;$fmt$,
			counter);
			
		EXECUTE FORMAT($fmt$CREATE TABLE test_table_%1$s 
			(
				field_block_A_1		int8,
				field_block_A_2		varchar,
				field_block_A_3		int8,
				field_block_A_4		int8,
				field_block_A_5		int8,
				field_block_A_6		varchar,
				field_block_A_7		timestamp,
				field_block_A_8		varchar,
				field_block_A_9		int8,
				field_block_A_10	numeric,
				field_block_A_11	int8,
				field_block_A_12	int4,
				field_block_A_13	numeric,
				field_block_A_14	int4,
				field_block_A_15	varchar,
				field_block_A_16	int4,
				field_block_A_17	int8,
				field_block_A_18	int4,
				field_block_A_19	varchar,
				field_block_A_20	int4,
				field_block_A_21	date,
				field_block_A_22	int4,
				field_block_A_23	int8,
				field_block_A_24	int4,
				field_block_A_25	varchar,
				field_block_A_26	int4,
				field_block_A_27	varchar,
				field_block_A_28	int4,
				field_block_A_29	numeric,
				field_block_A_30	int4,
				field_block_A_31	varchar,
				field_block_A_32	int4,
				field_block_A_33	varchar,
				field_block_A_34	int4,
				field_block_A_35	numeric,
				field_block_A_36	int4,
				field_block_A_37	varchar,
				field_block_A_38	int4,
				field_block_A_39	int4,
				field_block_A_40	int4,
				field_block_B_1		int8,
				field_block_B_2		varchar,
				field_block_B_3		int8,
				field_block_B_4		int8,
				field_block_B_5		int8,
				field_block_B_6		varchar,
				field_block_B_7		timestamp,
				field_block_B_8		varchar,
				field_block_B_9		int8,
				field_block_B_10	numeric,
				field_block_B_11	int8,
				field_block_B_12	int4,
				field_block_B_13	numeric,
				field_block_B_14	int4,
				field_block_B_15	varchar,
				field_block_B_16	int4,
				field_block_B_17	int8,
				field_block_B_18	int4,
				field_block_B_19	varchar,
				field_block_B_20	int4,
				field_block_B_21	date,
				field_block_B_22	int4,
				field_block_B_23	int8,
				field_block_B_24	int4,
				field_block_B_25	varchar,
				field_block_B_26	int4,
				field_block_B_27	varchar,
				field_block_B_28	int4,
				field_block_B_29	numeric,
				field_block_B_30	int4,
				field_block_B_31	varchar,
				field_block_B_32	int4,
				field_block_B_33	varchar,
				field_block_B_34	int4,
				field_block_B_35	numeric,
				field_block_B_36	int4,
				field_block_B_37	varchar,
				field_block_B_38	int4,
				field_block_B_39	int4,
				field_block_B_40	int4,
				field_block_C_1		int8,
				field_block_C_2		varchar,
				field_block_C_3		int8,
				field_block_C_4		int8,
				field_block_C_5		int8,
				field_block_C_6		varchar,
				field_block_C_7		timestamp,
				field_block_C_8		varchar,
				field_block_C_9		int8,
				field_block_C_10	numeric,
				field_block_C_11	int8,
				field_block_C_12	int4,
				field_block_C_13	numeric,
				field_block_C_14	int4,
				field_block_C_15	varchar,
				field_block_C_16	int4,
				field_block_C_17	int8,
				field_block_C_18	int4,
				field_block_C_19	varchar,
				field_block_C_20	int4,
				field_block_C_21	date,
				field_block_C_22	int4,
				field_block_C_23	int8,
				field_block_C_24	int4,
				field_block_C_25	varchar,
				field_block_C_26	int4,
				field_block_C_27	varchar,
				field_block_C_28	int4,
				field_block_C_29	numeric,
				field_block_C_30	int4,
				field_block_C_31	varchar,
				field_block_C_32	int4,
				field_block_C_33	varchar,
				field_block_C_34	int4,
				field_block_C_35	numeric,
				field_block_C_36	int4,
				field_block_C_37	varchar,
				field_block_C_38	int4,
				field_block_C_39	int4,
				field_block_C_40	int4
			) DISTRIBUTED BY (field_block_A_1);$fmt$,
			counter);
			
		EXECUTE FORMAT($fmt$INSERT INTO test_table_%1$s 
			VALUES
			(
			1, 'val_field_block_A_2', 3, 4, 5, 'val_field_block_A_6', '2021-01-01', 'val_field_block_A_8', 9, 10,
			11, 12, 13, 14, 'val_field_block_A_15', 16, 17, 18, 'val_field_block_A_19', 20, '2021-01-02', 22, 23,
			24, 'val_field_block_A_25', 26, 'val_field_block_A_27', 28, 29, 30, 'field_block_A_31', 32,
			'val_field_block_A_33', 34, 35, 36, 'val_field_block_A_37', 38, 39, 40,

			100, 'val_field_block_B_2', 103, 104, 105, 'val_field_block_B_6', '2021-02-01', 'val_field_block_B_8', 109, 110,
			111, 112, 113, 114, 'val_field_block_B_15', 116, 117, 118, 'val_field_block_B_19', 120, '2021-02-02', 122, 123,
			124, 'val_field_block_B_25', 126, 'val_field_block_B_27', 128, 129, 130, 'field_block_B_31', 132,
			'val_field_block_B_33', 134, 135, 136, 'val_field_block_B_37', 138, 139, 140,

			200, 'val_field_block_C_2', 203, 204, 205, 'val_field_block_C_6', '2021-03-01', 'val_field_block_C_8', 209, 210,
			211, 212, 213, 214, 'val_field_block_C_15', 216, 217, 218, 'val_field_block_C_19', 220, '2021-03-02', 222, 223,
			224, 'val_field_block_C_25', 226, 'val_field_block_C_27', 228, 229, 230, 'field_block_C_31', 232,
			'val_field_block_C_33', 234, 235, 236, 'val_field_block_C_37', 238, 239, 240
			),
			(
			1, 'val_field_block_A_2', 3, 4, 5, 'val_field_block_A_6', '2021-01-01', 'val_field_block_A_8', 9, 10,
			11, 12, 13, 14, 'val_field_block_A_15', 16, 17, 18, 'val_field_block_A_19', 20, '2021-01-02', 22, 23,
			24, 'val_field_block_A_25', 26, 'val_field_block_A_27', 28, 29, 30, 'field_block_A_31', 32,
			'val_field_block_A_33', 34, 35, 36, 'val_field_block_A_37', 38, 39, 40,

			100, 'val_field_block_B_2', 103, 104, 105, 'val_field_block_B_6', '2021-02-01', 'val_field_block_B_8', 109, 110,
			111, 112, 113, 114, 'val_field_block_B_15', 116, 117, 118, 'val_field_block_B_19', 120, '2021-02-02', 122, 123,
			124, 'val_field_block_B_25', 126, 'val_field_block_B_27', 128, 129, 130, 'field_block_B_31', 132,
			'val_field_block_B_33', 134, 135, 136, 'val_field_block_B_37', 138, 139, 140,

			200, 'val_field_block_C_2', 203, 204, 205, 'val_field_block_C_6', '2021-03-01', 'val_field_block_C_8', 209, 210,
			211, 212, 213, 214, 'val_field_block_C_15', 216, 217, 218, 'val_field_block_C_19', 220, '2021-03-02', 222, 223,
			224, 'val_field_block_C_25', 226, 'val_field_block_C_27', 228, 229, 230, 'field_block_C_31', 232,
			'val_field_block_C_33', 234, 235, 236, 'val_field_block_C_37', 238, 239, 240
			),
			(
			1, 'val_field_block_A_2', 3, 4, 5, 'val_field_block_A_6', '2021-01-01', 'val_field_block_A_8', 9, 10,
			11, 12, 13, 14, 'val_field_block_A_15', 16, 17, 18, 'val_field_block_A_19', 20, '2021-01-02', 22, 23,
			24, 'val_field_block_A_25', 26, 'val_field_block_A_27', 28, 29, 30, 'field_block_A_31', 32,
			'val_field_block_A_33', 34, 35, 36, 'val_field_block_A_37', 38, 39, 40,

			100, 'val_field_block_B_2', 103, 104, 105, 'val_field_block_B_6', '2021-02-01', 'val_field_block_B_8', 109, 110,
			111, 112, 113, 114, 'val_field_block_B_15', 116, 117, 118, 'val_field_block_B_19', 120, '2021-02-02', 122, 123,
			124, 'val_field_block_B_25', 126, 'val_field_block_B_27', 128, 129, 130, 'field_block_B_31', 132,
			'val_field_block_B_33', 134, 135, 136, 'val_field_block_B_37', 138, 139, 140,

			200, 'val_field_block_C_2', 203, 204, 205, 'val_field_block_C_6', '2021-03-01', 'val_field_block_C_8', 209, 210,
			211, 212, 213, 214, 'val_field_block_C_15', 216, 217, 218, 'val_field_block_C_19', 220, '2021-03-02', 222, 223,
			224, 'val_field_block_C_25', 226, 'val_field_block_C_27', 228, 229, 230, 'field_block_C_31', 232,
			'val_field_block_C_33', 234, 235, 236, 'val_field_block_C_37', 238, 239, 240
			)
			;$fmt$,
			counter);
	end loop;

END$$
LANGUAGE plpgsql VOLATILE
EXECUTE ON MASTER;
```

</details>